+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 --------------------------------------------------
+<a name="0.12.7"></a>
+# [0.12.7](https://github.com/moleculerjs/moleculer/compare/v0.12.6...v0.12.7) (2018-06-14)
+
+# Changes
+- fix action disabling with mixins [#298](https://github.com/moleculerjs/moleculer/issues/298)
+- Fix metrics options and add findNextActionEndpoint to index.d.ts
+
+--------------------------------------------------
 <a name="0.12.6"></a>
 # [0.12.6](https://github.com/moleculerjs/moleculer/compare/v0.12.5...v0.12.6) (2018-06-07)